pyproject.toml
[tool.poetry]
name = "conventional"
version = "0.5.1"
authors = ["David Symons <david@symons.io>"]
description = "No frills command-line tool for processing structured commits."

homepage = "https://github.com/multimac/conventional"
readme = "README.md"
license = "MIT"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Version Control :: Git",
]

[tool.poetry.dependencies]
aiocache = "^0.11.1"
colorama = "^0.4.3"
confuse = "^1.1.0"
jinja2 = "^2.11.2"
python = "^3.8"
python-dateutil = "^2.8.1"
typer = "^0.3"
flake8 = "^3.8.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^5.3"
mypy = "^0.782"
pytest = "^6.0"
pytest-asyncio = "^0.14.0"

[tool.poetry.scripts]
conventional = "conventional.cli:main"

[tool.isort]
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 99
multi_line_output = 3

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
